Move app contract schemas into new library
