+++ packages/schemas/README.md
@@ -1,0 +1,11 @@
+# schemas
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test schemas` to execute the unit tests via [Jest](https://jestjs.io).
+
+## Running lint
+
+Run `nx lint schemas` to execute the lint via [ESLint](https://eslint.org/).
