+++ packages/schemas/src/index.ts
@@ -1,0 +1,2 @@
+export * from './lib/api/routes';
+export * from './lib/shared';
