+++ packages/schemas/src/lib/api/routes/index.ts
@@ -1,0 +1,3 @@
+export * from './auth';
+export * from './import';
+export * from './playlists';
