+++ packages/schemas/src/lib/shared/content-type-schemas.ts
