+++ packages/schemas/src/lib/shared/index.ts
@@ -1,0 +1,2 @@
+export * from './content-type-schemas';
+export * from './spotify-token-schema';
